Simplify and unify PushPull-related logic (#1293)
Simplified document lookup logic by always returning a valid DocumentInfo.
Streamlined method signatures for document change and compaction processes.
Simplified struct initializations in tests to improve readability.
