+++ server/backend/database/testcases/testcases.go
@@ -57,19 +57,21 @@ func RunFindDocInfoTest(
 		clientInfo, err := db.ActivateClient(ctx, projectID, t.Name(), map[string]string{"userID": t.Name()})
 		assert.NoError(t, err)
 
-		_, err = db.FindDocInfoByRefKey(context.Background(), types.DocRefKey{
+		_, err = db.FindDocInfoByRefKey(ctx, types.DocRefKey{
 			ProjectID: projectID,
 			DocID:     dummyClientID,
 		})
 		assert.ErrorIs(t, err, database.ErrDocumentNotFound)
 
 		docKey := key.Key(fmt.Sprintf("tests$%s", t.Name()))
-		_, err = db.FindDocInfoByKeyAndOwner(ctx, clientInfo.RefKey(), docKey, false)
-		assert.ErrorIs(t, err, database.ErrDocumentNotFound)
-
-		docInfo, err := db.FindDocInfoByKeyAndOwner(ctx, clientInfo.RefKey(), docKey, true)
+		firstInfo, err := db.FindOrCreateDocInfo(ctx, clientInfo.RefKey(), docKey)
 		assert.NoError(t, err)
-		assert.Equal(t, docKey, docInfo.Key)
+		assert.Equal(t, docKey, firstInfo.Key)
+
+		secondInfo, err := db.FindOrCreateDocInfo(ctx, clientInfo.RefKey(), docKey)
+		assert.NoError(t, err)
+		assert.Equal(t, docKey, secondInfo.Key)
+		assert.Equal(t, firstInfo.ID, secondInfo.ID)
 	})
 
 	t.Run("find docInfo by key test", func(t *testing.T) {
@@ -79,12 +81,12 @@ func RunFindDocInfoTest(
 
 		// 01. Create a document
 		docKey := helper.TestDocKey(t)
-		info, err := db.FindDocInfoByKeyAndOwner(ctx, clientInfo.RefKey(), docKey, true)
+		info, err := db.FindOrCreateDocInfo(ctx, clientInfo.RefKey(), docKey)
 		assert.NoError(t, err)
 		assert.Equal(t, docKey, info.Key)
 
 		// 02. Remove the document
-		err = db.CreateChangeInfos(ctx, projectID, info, 0, []*change.Change{}, true)
+		err = db.CreateChangeInfos(ctx, info, 0, []*change.Change{}, true)
 		assert.NoError(t, err)
 
 		// 03. Find the document
@@ -110,7 +112,7 @@ func RunFindDocInfosByKeysTest(
 			"search$test", "abcde", "test abc",
 		}
 		for _, docKey := range docKeys {
-			_, err := db.FindDocInfoByKeyAndOwner(ctx, clientInfo.RefKey(), docKey, true)
+			_, err := db.FindOrCreateDocInfo(ctx, clientInfo.RefKey(), docKey)
 			assert.NoError(t, err)
 		}
 
@@ -138,7 +140,7 @@ func RunFindDocInfosByKeysTest(
 			"search$test", "abcde", "test abc",
 		}
 		for _, docKey := range docKeys {
-			_, err := db.FindDocInfoByKeyAndOwner(ctx, clientInfo.RefKey(), docKey, true)
+			_, err := db.FindOrCreateDocInfo(ctx, clientInfo.RefKey(), docKey)
 			assert.NoError(t, err)
 		}
 
@@ -158,7 +160,7 @@ func RunFindDocInfosByKeysTest(
 			"exist-key1", "exist-key2", "exist-key3",
 		}
 		for _, docKey := range docKeys {
-			_, err := db.FindDocInfoByKeyAndOwner(ctx, clientInfo.RefKey(), docKey, true)
+			_, err := db.FindOrCreateDocInfo(ctx, clientInfo.RefKey(), docKey)
 			assert.NoError(t, err)
 		}
 
@@ -272,7 +274,7 @@ func RunFindDocInfosByQueryTest(
 			"test0", "test1", "test2", "test3", "test10",
 			"test11", "test20", "test21", "test22", "test23"}
 		for _, docKey := range docKeys {
-			_, err := db.FindDocInfoByKeyAndOwner(ctx, clientInfo.RefKey(), key.Key(docKey), true)
+			_, err := db.FindOrCreateDocInfo(ctx, clientInfo.RefKey(), key.Key(docKey))
 			assert.NoError(t, err)
 		}
 
@@ -303,7 +305,7 @@ func RunFindChangesBetweenServerSeqsTest(
 		docKey := key.Key(fmt.Sprintf("tests$%s", t.Name()))
 
 		clientInfo, _ := db.ActivateClient(ctx, projectID, t.Name(), map[string]string{"userID": t.Name()})
-		docInfo, _ := db.FindDocInfoByKeyAndOwner(ctx, clientInfo.RefKey(), docKey, true)
+		docInfo, _ := db.FindOrCreateDocInfo(ctx, clientInfo.RefKey(), docKey)
 		docRefKey := docInfo.RefKey()
 		assert.NoError(t, clientInfo.AttachDocument(docInfo.ID, false))
 		assert.NoError(t, db.UpdateClientInfoAfterPushPull(ctx, clientInfo, docInfo))
@@ -328,7 +330,7 @@ func RunFindChangesBetweenServerSeqsTest(
 		}
 
 		// Store changes
-		err := db.CreateChangeInfos(ctx, projectID, docInfo, 0, pack.Changes, false)
+		err := db.CreateChangeInfos(ctx, docInfo, 0, pack.Changes, false)
 		assert.NoError(t, err)
 
 		// Find changes
@@ -355,7 +357,7 @@ func RunFindChangeInfosBetweenServerSeqsTest(
 		docKey := key.Key(fmt.Sprintf("tests$%s", t.Name()))
 
 		clientInfo, _ := db.ActivateClient(ctx, projectID, t.Name(), map[string]string{"userID": t.Name()})
-		docInfo, _ := db.FindDocInfoByKeyAndOwner(ctx, clientInfo.RefKey(), docKey, true)
+		docInfo, _ := db.FindOrCreateDocInfo(ctx, clientInfo.RefKey(), docKey)
 		assert.NoError(t, clientInfo.AttachDocument(docInfo.ID, false))
 		assert.NoError(t, db.UpdateClientInfoAfterPushPull(ctx, clientInfo, docInfo))
 
@@ -385,7 +387,7 @@ func RunFindChangeInfosBetweenServerSeqsTest(
 		docKey := key.Key(fmt.Sprintf("tests$%s", t.Name()))
 
 		clientInfo, _ := db.ActivateClient(ctx, projectID, t.Name(), map[string]string{"userID": t.Name()})
-		docInfo, _ := db.FindDocInfoByKeyAndOwner(ctx, clientInfo.RefKey(), docKey, true)
+		docInfo, _ := db.FindOrCreateDocInfo(ctx, clientInfo.RefKey(), docKey)
 		docRefKey := docInfo.RefKey()
 		assert.NoError(t, clientInfo.AttachDocument(docInfo.ID, false))
 		assert.NoError(t, db.UpdateClientInfoAfterPushPull(ctx, clientInfo, docInfo))
@@ -416,7 +418,6 @@ func RunFindChangeInfosBetweenServerSeqsTest(
 
 		err := db.CreateChangeInfos(
 			ctx,
-			projectID,
 			docInfo,
 			initialServerSeq,
 			pack.Changes,
@@ -464,7 +465,7 @@ func RunFindChangeInfosBetweenServerSeqsTest(
 		docKey := key.Key(fmt.Sprintf("tests$%s", t.Name()))
 
 		clientInfo, _ := db.ActivateClient(ctx, projectID, t.Name(), map[string]string{"userID": t.Name()})
-		docInfo, _ := db.FindDocInfoByKeyAndOwner(ctx, clientInfo.RefKey(), docKey, true)
+		docInfo, _ := db.FindOrCreateDocInfo(ctx, clientInfo.RefKey(), docKey)
 		docRefKey := docInfo.RefKey()
 		assert.NoError(t, clientInfo.AttachDocument(docInfo.ID, false))
 		assert.NoError(t, db.UpdateClientInfoAfterPushPull(ctx, clientInfo, docInfo))
@@ -494,7 +495,6 @@ func RunFindChangeInfosBetweenServerSeqsTest(
 
 		err := db.CreateChangeInfos(
 			ctx,
-			projectID,
 			docInfo,
 			initialServerSeq,
 			pack.Changes,
@@ -536,7 +536,7 @@ func RunFindLatestChangeInfoTest(t *testing.T,
 		// 01. Activate client and find document info.
 		clientInfo, err := db.ActivateClient(ctx, projectID, t.Name(), map[string]string{"userID": t.Name()})
 		assert.NoError(t, err)
-		docInfo, err := db.FindDocInfoByKeyAndOwner(ctx, clientInfo.RefKey(), docKey, true)
+		docInfo, err := db.FindOrCreateDocInfo(ctx, clientInfo.RefKey(), docKey)
 		assert.NoError(t, err)
 		docRefKey := docInfo.RefKey()
 		assert.NoError(t, clientInfo.AttachDocument(docInfo.ID, false))
@@ -570,7 +570,6 @@ func RunFindLatestChangeInfoTest(t *testing.T,
 
 		assert.NoError(t, db.CreateChangeInfos(
 			ctx,
-			projectID,
 			docInfo,
 			initialServerSeq,
 			pack.Changes,
@@ -608,7 +607,7 @@ func RunFindClosestSnapshotInfoTest(t *testing.T, db database.Database, projectI
 		clientInfo, _ := db.ActivateClient(ctx, projectID, t.Name(), map[string]string{"userID": t.Name()})
 		bytesID, _ := clientInfo.ID.Bytes()
 		actorID, _ := time.ActorIDFromBytes(bytesID)
-		docInfo, _ := db.FindDocInfoByKeyAndOwner(ctx, clientInfo.RefKey(), docKey, true)
+		docInfo, _ := db.FindOrCreateDocInfo(ctx, clientInfo.RefKey(), docKey)
 
 		doc := document.New(key.Key(t.Name()))
 		doc.SetActor(actorID)
@@ -896,7 +895,7 @@ func RunFindDocInfosByPagingTest(t *testing.T, db database.Database, projectID t
 		clientInfo, _ := db.ActivateClient(ctx, projectID, t.Name(), map[string]string{"userID": t.Name()})
 		docInfos := make([]*database.DocInfo, 0, totalSize)
 		for i := 0; i < totalSize; i++ {
-			docInfo, err := db.FindDocInfoByKeyAndOwner(ctx, clientInfo.RefKey(), key.Key(fmt.Sprintf("%d", i)), true)
+			docInfo, err := db.FindOrCreateDocInfo(ctx, clientInfo.RefKey(), key.Key(fmt.Sprintf("%d", i)))
 			assert.NoError(t, err)
 			docInfos = append(docInfos, docInfo)
 		}
@@ -960,10 +959,10 @@ func RunFindDocInfosByPagingTest(t *testing.T, db database.Database, projectID t
 		var dummyDocInfos []*database.DocInfo
 		for i := 0; i <= testDocCnt; i++ {
 			testDocKey := key.Key(fmt.Sprintf("%s%02d", "testdockey", i))
-			docInfo, err := db.FindDocInfoByKeyAndOwner(ctx, types.ClientRefKey{
+			docInfo, err := db.FindOrCreateDocInfo(ctx, types.ClientRefKey{
 				ProjectID: testProjectInfo.ID,
 				ClientID:  dummyClientID,
-			}, testDocKey, true)
+			}, testDocKey)
 			assert.NoError(t, err)
 			dummyDocInfos = append(dummyDocInfos, docInfo)
 		}
@@ -1066,10 +1065,10 @@ func RunFindDocInfosByPagingTest(t *testing.T, db database.Database, projectID t
 		var docInfos []*database.DocInfo
 		for i := 0; i < testDocCnt; i++ {
 			testDocKey := key.Key("key" + strconv.Itoa(i))
-			docInfo, err := db.FindDocInfoByKeyAndOwner(ctx, types.ClientRefKey{
+			docInfo, err := db.FindOrCreateDocInfo(ctx, types.ClientRefKey{
 				ProjectID: projectInfo.ID,
 				ClientID:  dummyClientID,
-			}, testDocKey, true)
+			}, testDocKey)
 			assert.NoError(t, err)
 			docInfos = append(docInfos, docInfo)
 		}
@@ -1089,9 +1088,9 @@ func RunFindDocInfosByPagingTest(t *testing.T, db database.Database, projectID t
 		AssertKeys(t, docKeysInReverse, result)
 
 		// 03. Remove some documents.
-		err = db.CreateChangeInfos(ctx, projectInfo.ID, docInfos[1], 0, []*change.Change{}, true)
+		err = db.CreateChangeInfos(ctx, docInfos[1], 0, []*change.Change{}, true)
 		assert.NoError(t, err)
-		err = db.CreateChangeInfos(ctx, projectInfo.ID, docInfos[3], 0, []*change.Change{}, true)
+		err = db.CreateChangeInfos(ctx, docInfos[3], 0, []*change.Change{}, true)
 		assert.NoError(t, err)
 
 		// 04. List the documents again and check the filtered result.
@@ -1113,13 +1112,13 @@ func RunCreateChangeInfosTest(t *testing.T, db database.Database, projectID type
 
 		// 01. Create a client and a document then attach the document to the client.
 		clientInfo, _ := db.ActivateClient(ctx, projectID, t.Name(), map[string]string{"userID": t.Name()})
-		docInfo, _ := db.FindDocInfoByKeyAndOwner(ctx, clientInfo.RefKey(), docKey, true)
+		docInfo, _ := db.FindOrCreateDocInfo(ctx, clientInfo.RefKey(), docKey)
 		docRefKey := docInfo.RefKey()
 		assert.NoError(t, clientInfo.AttachDocument(docInfo.ID, false))
 		assert.NoError(t, db.UpdateClientInfoAfterPushPull(ctx, clientInfo, docInfo))
 
 		// 02. Remove the document and check the document is removed.
-		err := db.CreateChangeInfos(ctx, projectID, docInfo, 0, []*change.Change{}, true)
+		err := db.CreateChangeInfos(ctx, docInfo, 0, []*change.Change{}, true)
 		assert.NoError(t, err)
 		docInfo, err = db.FindDocInfoByRefKey(ctx, docRefKey)
 		assert.NoError(t, err)
@@ -1132,18 +1131,18 @@ func RunCreateChangeInfosTest(t *testing.T, db database.Database, projectID type
 
 		// 01. Create a client and a document then attach the document to the client.
 		clientInfo1, _ := db.ActivateClient(ctx, projectID, t.Name(), map[string]string{"userID": t.Name()})
-		docInfo1, _ := db.FindDocInfoByKeyAndOwner(ctx, clientInfo1.RefKey(), docKey, true)
+		docInfo1, _ := db.FindOrCreateDocInfo(ctx, clientInfo1.RefKey(), docKey)
 		docRefKey1 := docInfo1.RefKey()
 		assert.NoError(t, clientInfo1.AttachDocument(docRefKey1.DocID, false))
 		assert.NoError(t, db.UpdateClientInfoAfterPushPull(ctx, clientInfo1, docInfo1))
 
 		// 02. Remove the document.
 		assert.NoError(t, clientInfo1.RemoveDocument(docRefKey1.DocID))
-		err := db.CreateChangeInfos(ctx, projectID, docInfo1, 0, []*change.Change{}, true)
+		err := db.CreateChangeInfos(ctx, docInfo1, 0, []*change.Change{}, true)
 		assert.NoError(t, err)
 
 		// 03. Create a document with same key and check they have same key but different id.
-		docInfo2, _ := db.FindDocInfoByKeyAndOwner(ctx, clientInfo1.RefKey(), docKey, true)
+		docInfo2, _ := db.FindOrCreateDocInfo(ctx, clientInfo1.RefKey(), docKey)
 		docRefKey2 := docInfo2.RefKey()
 		assert.NoError(t, clientInfo1.AttachDocument(docRefKey2.DocID, false))
 		assert.NoError(t, db.UpdateClientInfoAfterPushPull(ctx, clientInfo1, docInfo2))
@@ -1156,7 +1155,7 @@ func RunCreateChangeInfosTest(t *testing.T, db database.Database, projectID type
 		docKey := key.Key(fmt.Sprintf("tests$%s", t.Name()))
 
 		clientInfo, _ := db.ActivateClient(ctx, projectID, t.Name(), map[string]string{"userID": t.Name()})
-		docInfo, _ := db.FindDocInfoByKeyAndOwner(ctx, clientInfo.RefKey(), docKey, true)
+		docInfo, _ := db.FindOrCreateDocInfo(ctx, clientInfo.RefKey(), docKey)
 		docRefKey := docInfo.RefKey()
 		assert.NoError(t, clientInfo.AttachDocument(docInfo.ID, false))
 		assert.NoError(t, db.UpdateClientInfoAfterPushPull(ctx, clientInfo, docInfo))
@@ -1166,7 +1165,7 @@ func RunCreateChangeInfosTest(t *testing.T, db database.Database, projectID type
 
 		// Set removed_at in docInfo and store changes
 		assert.NoError(t, clientInfo.RemoveDocument(docInfo.ID))
-		err := db.CreateChangeInfos(ctx, projectID, docInfo, 0, pack.Changes, true)
+		err := db.CreateChangeInfos(ctx, docInfo, 0, pack.Changes, true)
 		assert.NoError(t, err)
 
 		// Check whether removed_at is set in docInfo
@@ -1186,7 +1185,7 @@ func RunCreateChangeInfosTest(t *testing.T, db database.Database, projectID type
 
 		// 01. Create a client and a document then attach the document to the client.
 		clientInfo, _ := db.ActivateClient(ctx, projectID, t.Name(), map[string]string{"userID": t.Name()})
-		docInfo1, _ := db.FindDocInfoByKeyAndOwner(ctx, clientInfo.RefKey(), docKey, true)
+		docInfo1, _ := db.FindOrCreateDocInfo(ctx, clientInfo.RefKey(), docKey)
 		assert.Equal(t, docInfo1.Owner, clientInfo.ID)
 		assert.NotEqual(t, gotime.Date(1, gotime.January, 1, 0, 0, 0, 0, gotime.UTC), docInfo1.UpdatedAt)
 		assert.Equal(t, docInfo1.CreatedAt, docInfo1.UpdatedAt)
@@ -1206,8 +1205,8 @@ func RunCreateChangeInfosTest(t *testing.T, db database.Database, projectID type
 		}))
 		pack := doc.CreateChangePack()
 		updatedAt := docInfo1.UpdatedAt
-		assert.NoError(t, db.CreateChangeInfos(ctx, projectID, docInfo1, 0, pack.Changes, false))
-		docInfo2, _ := db.FindDocInfoByKeyAndOwner(ctx, clientInfo.RefKey(), docKey, true)
+		assert.NoError(t, db.CreateChangeInfos(ctx, docInfo1, 0, pack.Changes, false))
+		docInfo2, _ := db.FindOrCreateDocInfo(ctx, clientInfo.RefKey(), docKey)
 		assert.Equal(t, updatedAt, docInfo2.UpdatedAt)
 
 		// 03. Update document presence and operation
@@ -1218,8 +1217,8 @@ func RunCreateChangeInfosTest(t *testing.T, db database.Database, projectID type
 		}))
 		pack = doc.CreateChangePack()
 		updatedAt = docInfo2.UpdatedAt
-		assert.NoError(t, db.CreateChangeInfos(ctx, projectID, docInfo2, 0, pack.Changes, false))
-		docInfo3, _ := db.FindDocInfoByKeyAndOwner(ctx, clientInfo.RefKey(), docKey, true)
+		assert.NoError(t, db.CreateChangeInfos(ctx, docInfo2, 0, pack.Changes, false))
+		docInfo3, _ := db.FindOrCreateDocInfo(ctx, clientInfo.RefKey(), docKey)
 		assert.NotEqual(t, updatedAt, docInfo3.UpdatedAt)
 	})
 }
@@ -1234,7 +1233,7 @@ func RunUpdateClientInfoAfterPushPullTest(t *testing.T, db database.Database, pr
 		assert.NoError(t, err)
 
 		docKey := key.Key(fmt.Sprintf("tests$%s", t.Name()))
-		docInfo, err := db.FindDocInfoByKeyAndOwner(ctx, clientInfo.RefKey(), docKey, true)
+		docInfo, err := db.FindOrCreateDocInfo(ctx, clientInfo.RefKey(), docKey)
 		assert.NoError(t, err)
 
 		err = db.UpdateClientInfoAfterPushPull(ctx, clientInfo, docInfo)
@@ -1248,7 +1247,7 @@ func RunUpdateClientInfoAfterPushPullTest(t *testing.T, db database.Database, pr
 		assert.NoError(t, err)
 
 		docKey := key.Key(fmt.Sprintf("tests$%s", t.Name()))
-		docInfo, err := db.FindDocInfoByKeyAndOwner(ctx, clientInfo.RefKey(), docKey, true)
+		docInfo, err := db.FindOrCreateDocInfo(ctx, clientInfo.RefKey(), docKey)
 		assert.NoError(t, err)
 
 		assert.NoError(t, clientInfo.AttachDocument(docInfo.ID, false))
@@ -1266,7 +1265,7 @@ func RunUpdateClientInfoAfterPushPullTest(t *testing.T, db database.Database, pr
 		assert.NoError(t, err)
 
 		docKey := key.Key(fmt.Sprintf("tests$%s", t.Name()))
-		docInfo, err := db.FindDocInfoByKeyAndOwner(ctx, clientInfo.RefKey(), docKey, true)
+		docInfo, err := db.FindOrCreateDocInfo(ctx, clientInfo.RefKey(), docKey)
 		assert.NoError(t, err)
 
 		assert.NoError(t, clientInfo.AttachDocument(docInfo.ID, false))
@@ -1308,7 +1307,7 @@ func RunUpdateClientInfoAfterPushPullTest(t *testing.T, db database.Database, pr
 		assert.NoError(t, err)
 
 		docKey := key.Key(fmt.Sprintf("tests$%s", t.Name()))
-		docInfo, err := db.FindDocInfoByKeyAndOwner(ctx, clientInfo.RefKey(), docKey, true)
+		docInfo, err := db.FindOrCreateDocInfo(ctx, clientInfo.RefKey(), docKey)
 		assert.NoError(t, err)
 
 		assert.NoError(t, clientInfo.AttachDocument(docInfo.ID, false))
@@ -1337,7 +1336,7 @@ func RunUpdateClientInfoAfterPushPullTest(t *testing.T, db database.Database, pr
 		assert.NoError(t, err)
 
 		docKey := key.Key(fmt.Sprintf("tests$%s", t.Name()))
-		docInfo, err := db.FindDocInfoByKeyAndOwner(ctx, clientInfo.RefKey(), docKey, true)
+		docInfo, err := db.FindOrCreateDocInfo(ctx, clientInfo.RefKey(), docKey)
 		assert.NoError(t, err)
 
 		assert.NoError(t, clientInfo.AttachDocument(docInfo.ID, false))
@@ -1366,7 +1365,7 @@ func RunUpdateClientInfoAfterPushPullTest(t *testing.T, db database.Database, pr
 		assert.NoError(t, err)
 
 		docKey := key.Key(fmt.Sprintf("tests$%s", t.Name()))
-		docInfo, err := db.FindDocInfoByKeyAndOwner(ctx, clientInfo.RefKey(), docKey, true)
+		docInfo, err := db.FindOrCreateDocInfo(ctx, clientInfo.RefKey(), docKey)
 		assert.NoError(t, err)
 
 		assert.NoError(t, clientInfo.AttachDocument(docInfo.ID, false))
@@ -1390,7 +1389,7 @@ func RunIsDocumentAttachedTest(t *testing.T, db database.Database, projectID typ
 		assert.NoError(t, err)
 		c2, err := db.ActivateClient(ctx, projectID, t.Name()+"2", map[string]string{"userID": t.Name() + "2"})
 		assert.NoError(t, err)
-		d1, err := db.FindDocInfoByKeyAndOwner(ctx, c1.RefKey(), helper.TestDocKey(t), true)
+		d1, err := db.FindOrCreateDocInfo(ctx, c1.RefKey(), helper.TestDocKey(t))
 		assert.NoError(t, err)
 
 		// 01. Check if document is attached without attaching
@@ -1443,9 +1442,9 @@ func RunIsDocumentAttachedTest(t *testing.T, db database.Database, projectID typ
 		// 00. Create a client and two documents
 		c1, err := db.ActivateClient(ctx, projectID, t.Name()+"1", map[string]string{"userID": t.Name() + "1"})
 		assert.NoError(t, err)
-		d1, err := db.FindDocInfoByKeyAndOwner(ctx, c1.RefKey(), helper.TestDocKey(t)+"1", true)
+		d1, err := db.FindOrCreateDocInfo(ctx, c1.RefKey(), helper.TestDocKey(t)+"1")
 		assert.NoError(t, err)
-		d2, err := db.FindDocInfoByKeyAndOwner(ctx, c1.RefKey(), helper.TestDocKey(t)+"2", true)
+		d2, err := db.FindOrCreateDocInfo(ctx, c1.RefKey(), helper.TestDocKey(t)+"2")
 		assert.NoError(t, err)
 
 		// 01. Check if documents are attached after attaching
@@ -1489,7 +1488,7 @@ func RunIsDocumentAttachedTest(t *testing.T, db database.Database, projectID typ
 		assert.NoError(t, err)
 		c2, err := db.ActivateClient(ctx, projectID, t.Name()+"2", map[string]string{"userID": t.Name() + "2"})
 		assert.NoError(t, err)
-		d1, err := db.FindDocInfoByKeyAndOwner(ctx, c1.RefKey(), helper.TestDocKey(t), true)
+		d1, err := db.FindOrCreateDocInfo(ctx, c1.RefKey(), helper.TestDocKey(t))
 		assert.NoError(t, err)
 
 		// 01. Check if document is attached without attaching
@@ -1652,7 +1651,7 @@ func RunFindClientInfosByAttachedDocRefKeyTest(t *testing.T, db database.Databas
 		docKey := key.Key(fmt.Sprintf("tests$%s", t.Name()))
 
 		clientInfo1, _ := db.ActivateClient(ctx, projectID, t.Name(), map[string]string{"userID": t.Name()})
-		docInfo, _ := db.FindDocInfoByKeyAndOwner(ctx, clientInfo1.RefKey(), docKey, true)
+		docInfo, _ := db.FindOrCreateDocInfo(ctx, clientInfo1.RefKey(), docKey)
 		docRefKey := docInfo.RefKey()
 		assert.NoError(t, clientInfo1.AttachDocument(docRefKey.DocID, false))
 		assert.NoError(t, db.UpdateClientInfoAfterPushPull(ctx, clientInfo1, docInfo))
@@ -1694,7 +1693,7 @@ func RunPurgeDocument(t *testing.T, db database.Database, projectID types.ID) {
 		// 01. Create a client and a document then attach the document to the client.
 		clientInfo, _ := db.ActivateClient(ctx, projectID, t.Name(), map[string]string{"userID": t.Name()})
 		docKey := key.Key(fmt.Sprintf("tests$%s", t.Name()))
-		docInfo, _ := db.FindDocInfoByKeyAndOwner(ctx, clientInfo.RefKey(), docKey, true)
+		docInfo, _ := db.FindOrCreateDocInfo(ctx, clientInfo.RefKey(), docKey)
 		docRefKey := docInfo.RefKey()
 		assert.NoError(t, clientInfo.AttachDocument(docRefKey.DocID, false))
 		assert.NoError(t, db.UpdateClientInfoAfterPushPull(ctx, clientInfo, docInfo))
