+++ test/complex/mongo_client_test.go
@@ -136,10 +136,10 @@ func TestClientWithShardedDB(t *testing.T) {
 		docKey2 := key.Key(fmt.Sprintf("%s%d", t.Name(), 2))
 
 		// 01. Initialize a project and create a document.
-		docInfo1, err := cli.FindDocInfoByKeyAndOwner(ctx, types.ClientRefKey{
+		docInfo1, err := cli.FindOrCreateDocInfo(ctx, types.ClientRefKey{
 			ProjectID: projectID1,
 			ClientID:  dummyClientID,
-		}, docKey1, true)
+		}, docKey1)
 		assert.NoError(t, err)
 
 		// 02. Create an extra document with duplicate ID.
